+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.6-pre
 * Add parameter support for test commands
 * Cmdt no longer breaks when YAML file is empty
+* Fix test command execution error propagation
 
 ### 0.0.5
 * Fix sample test file on README.